[Fix #2323] Fix skull ticks overflow
- Changed skull ticks from millisecond to seconds to avoid overflow and
preserve performance because there's no need for using milliseconds.

Fixes[#2323]

+++ schema.sql
@@ -44,7 +44,7 @@ CREATE TABLE IF NOT EXISTS `players` (
   `lastip` int(10) unsigned NOT NULL DEFAULT '0',
   `save` tinyint(1) NOT NULL DEFAULT '1',
   `skull` tinyint(1) NOT NULL DEFAULT '0',
-  `skulltime` int(11) NOT NULL DEFAULT '0',
+  `skulltime` bigint(20) NOT NULL DEFAULT '0',
   `lastlogout` bigint(20) unsigned NOT NULL DEFAULT '0',
   `blessings` tinyint(2) NOT NULL DEFAULT '0',
   `onlinetime` int(11) NOT NULL DEFAULT '0',
